requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Optional dependencies for future implementation
beautifulsoup4>=4.12.0
requests>=2.31.0
